refactored routes into individual files, with dynamic handling
