+++ public/forum.html
@@ -1,0 +1,149 @@
+<!DOCTYPE html>
+<html lang="en">
+	<head>
+		<meta charset="UTF-8">
+		<meta name="viewport" content="width=device-width, initial-scale=1.0">
+		<title>CloudClub</title>
+
+		<!--data table-->
+		<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
+
+		<script type='text/javascript' src='https://cdn.datatables.net/1.10.23/js/jquery.dataTables.min.js'></script>
+		<script type='text/javascript' src='https://cdn.datatables.net/1.10.23/js/dataTables.bootstrap4.min.js'></script>
+		<link rel='stylesheet' type='text/css' href='https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.5.2/css/bootstrap.css'>
+		<link rel='stylesheet' type='text/css' href='https://cdn.datatables.net/1.10.23/css/dataTables.bootstrap4.min.css'>
+	</head>
+
+	<script>
+		function openInNewTab(url) {
+		  var win = window.open(url, '_blank');
+		  win.focus();
+		}
+	</script>
+
+	<style>
+		#project-stab { 
+			padding-left: 15px;
+			padding-right: 15px;
+		}
+	</style>
+
+	<body>
+		<!-- HEADER CODE -->
+		<header>
+            <!-- Note: changed header from fixed top to sticky-top on this page.
+            Ideally, header is only fixed on the homepage (to fit the carousel nicely with viewport sizing) -->
+			<nav class="navbar navbar-expand-md navbar-dark sticky-top bg-dark">
+				<a class="navbar-brand" href="/">CloudClub</a>
+				<button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarCollapse"
+					aria-controls="navbarCollapse" aria-expanded="false" aria-label="Toggle navigation">
+					<span class="navbar-toggler-icon"></span>
+				</button>
+				<div class="collapse navbar-collapse" id="navbarCollapse">
+					<ul class="navbar-nav mr-auto">
+						<li class="nav-item">
+							<a class="nav-link" href="/">Homepage <span class="sr-only">(current)</span></a>
+						</li>
+						<li class="nav-item">
+							<a class="nav-link" href="teampage.html">The Team</a>
+						</li>
+						<li class="nav-item">
+							<a class="nav-link" href="projects.html">Projects</a>
+                        </li>
+                        <li class="nav-item active">
+                            <a class="nav-link" href="forum.html">Forum</a>
+						</li>	
+					
+						<li class="nav-item">
+							<a class="nav-link" href="" onclick='openInNewTab("https://github.com/CloudClub-uoft");''>GitHub</a>
+						</li>
+					</ul>
+				</div>
+			</nav>
+        </header>
+
+		<!-- Heading and Description of Forum Page -->
+		<div class="list-group"></div>
+		<a href="#" class="list-group-item list-group-item-action active">
+			<div class="d-flex w-100 justify-content-between">
+				<div class="forum-title">
+					<h1 class="text-white">CloudConvos ☁️</h1>
+					<p class="text-white">
+						This name is temporary don't worry.
+					</p>
+				</div>
+			</div>
+		</a>
+		
+		<br>
+		<script>
+			$(document).ready( function () {
+				$('#table_id').DataTable({
+					"pagingType": "full_numbers",
+					"stateSave": true,
+				});
+			} );
+		</script>
+		<div class="container">
+			<div class="col">
+			<table id="table_id" class="table">
+				<thead>
+					<tr>
+						<th>Posts</th>
+					</tr>
+				</thead>
+				<tbody>
+					<tr>
+
+						<!-- Test post format. Could use some padding. -->
+						<td> <a href="#" class="list-group-item list-group-item-action">
+							<div class="d-flex w-100 justify-content-between">
+								<div class="d-flex w-80">
+									<div class="d-flex flex-column justify-content-between">
+										<h5 class="mb-1">1 - Post name</h5>
+										<p class="mb-1">Post description. Donec id elit non mi porta gravida at eget metus. Maecenas sed diam eget risus varius blandit.</p>
+										<small class="text-muted">Posted by Username</small>
+									</div>
+								</div>
+								<div class="d-flex w-30 justify-content-end">
+									<div class="d-flex flex-column justify-content-between">
+										<small class="text-muted">3 days ago</small>
+										<div class="d-flex flex-column justify-content-between">
+											<small class="text-muted">10 views</small>
+											<small class="text-muted">30 replies</small>
+										</div>
+									</div>
+								</div>
+							</div>
+						</a></td>
+		
+					</tr>
+					<tr>
+						<td> <a href="#" class="list-group-item list-group-item-action">
+							<div class="d-flex w-100 justify-content-between">
+							  <h5 class="mb-1">2-List group item heading</h5>
+							  <small class="text-muted">3 days ago</small>
+							</div>
+							<p class="mb-1">Post description. Donec id elit non mi porta gravida at eget metus. Maecenas sed diam eget risus varius blandit.</p>
+							<small class="text-muted">Donec id elit non mi porta.</small>
+						  </a></td>
+				
+					</tr>
+					<tr>
+						<td> <a href="#" class="list-group-item list-group-item-action">
+							<div class="d-flex w-100 justify-content-between">
+							  <h5 class="mb-1">2-List group item heading</h5>
+							  <small class="text-muted">3 days ago</small>
+							</div>
+							<p class="mb-1">Post description. Donec id elit non mi porta gravida at eget metus. Maecenas sed diam eget risus varius blandit.</p>
+							<small class="text-muted">Donec id elit non mi porta.</small>
+						  </a></td>
+				
+					</tr>
+				</tbody>
+			</table>
+		</div>
+		</div>
+	</body>
+
+</html>
