+++ public/projects.html
@@ -1,6 +1,5 @@
 <!DOCTYPE html>
 <html lang="en">
-
 	<head>
 		<meta charset="UTF-8">
 		<meta name="viewport" content="width=device-width, initial-scale=1.0">
@@ -41,7 +40,7 @@
 
 		<!-- HEADER CODE -->
 		<header>
-			<nav class="navbar navbar-expand-md navbar-dark fixed-top bg-dark">
+			<nav class="navbar navbar-expand-md navbar-dark sticky-top bg-dark">
 				<a class="navbar-brand" href="/">CloudClub</a>
 				<button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarCollapse"
 					aria-controls="navbarCollapse" aria-expanded="false" aria-label="Toggle navigation">
@@ -57,7 +56,10 @@
 						</li>
 						<li class="nav-item active">
 							<a class="nav-link" href="projects.html">Projects</a>
-						</li>						
+						</li>			
+						<li class="nav-item">
+							<a class="nav-link" href="forum.html">Forum</a>
+						</li>			
 						<li class="nav-item">
 							<a class="nav-link" href="" onclick='openInNewTab("https://github.com/CloudClub-uoft");''>GitHub</a>
 						</li>
@@ -106,6 +108,7 @@
 					<div class="p-5"> <a href="" class="btn btn-primary my-auto">Coming Soon!</a></div>
 				</div>
 			</div>
+		</div>
 
 		<!-- Test Footer -->
 		<footer class="container py-5 text-center" >
