+++ public/teampage.html
@@ -44,6 +44,9 @@
             <a class="nav-link" href="projects.html" tabindex="-1" aria-disabled="true">Projects</a>
           </li>
           <li class="nav-item">
+            <a class="nav-link" href="forum.html">Forum</a>
+          </li>
+          <li class="nav-item">
             <a class="nav-link" href="" onclick='openInNewTab("https:\/\/github.com/CloudClub-uoft");'>GitHub</a>
           </li>
         </ul>
@@ -419,7 +422,7 @@
           <div class="btn-group">
             <!-- <a role="a" class="btn btn-sm btn-info" href="https://github.com/Tamim1236"
             target="_blank">GitHub</a> -->
-            <a role="a" class="btn btn-sm btn-primary" href="www.linkedin/in/yanfei-zhu-cn"
+            <a role="a" class="btn btn-sm btn-primary" href="http://www.linkedin.com/in/yanfei-zhu-cn"
             target="_blank">LinkedIn</a>
           </div>
         </div>
