+++ routes/GET-index.js
@@ -1,0 +1,9 @@
+module.exports = (app) => {
+  app.get('/', (req, res) => {
+    const sesh = req.session;
+    if (sesh.email) {
+      return res.render('index', { email: sesh.email });
+    }
+    return res.render('index', { email: false });
+  });
+};
