+++ routes/GET-login.js
@@ -1,0 +1,5 @@
+module.exports = (app) => {
+  app.get('/login', (req, res) => {
+    res.render('login');
+  });
+};
