+++ routes/GET-register.js
@@ -1,0 +1,5 @@
+module.exports = (app) => {
+  app.get('/register', (req, res) => {
+    res.render('register');
+  });
+};
