+++ views/index.ejs
@@ -54,6 +54,9 @@
           <a class="nav-link" href="projects.html" tabindex="-1" aria-disabled="true">Projects</a>
         </li>
         <li class="nav-item">
+          <a class="nav-link" href="forum.html">Forum</a>
+        </li>
+        <li class="nav-item">
           <a class="nav-link" href="" onclick='openInNewTab("https:\/\/github.com/CloudClub-uoft");'>GitHub</a>
         </li>
       </ul>
